pyproject.toml
[project]
name = "Lizmap-server-plugin"
authors = [
    { name = "3liz", email = "info@3liz.com" } 
]
requires-python = ">=3.10"
keywords = ["QGIS", "processing", "plugin"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: GIS",
]
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/3liz/qgis-edigeo-processing"
repository = "https://github.com/3liz/qgis-edigeo-processing"


[dependency-groups]
dev = [
    "coverage[toml]",
    {include-group = "tests"},
    {include-group = "lint"},
]
tests = [
    "pytest",
    "pytest-qgis",
    "semver",
]
lint = [
    "ruff",
    "mypy",
]
packaging = [
    "qgis-plugin-package-ci; python_full_version >= '3.12'",
]


[tool.uv.sources]
qgis-plugin-package-ci = { git = "https://github.com/3liz/qgis-plugin-package-ci.git" }

#
# Package ci
#

[tool.package-ci]
plugin_source = "qgis_edigeo_processing"
github_organization_slug = "3liz"
project_slug = "qgis-edigeo-processing"

#
# Ruff
#
# https://doc.astral.sh/ruff/configuration
#

[tool.ruff]
line-length = 110
target-version = "py312"
exclude = [
    ".venv",
    ".local",
    "tests/fixtures",
]

[tool.ruff.format]
indent-style = "space"

[tool.ruff.lint]
extend-select = [
    "E", "F", "I", "ANN", "W", "T", "COM", "RUF",
    "C4", "SIM", "TC", "RET",
]
# COM812 conflict with formatter
ignore = [
    "ANN002", 
    "ANN003",
    "ANN401", # Dynamically typed expressions (Any)
    "RUF100", 
    "RUF029", # Unused async,
    "RET505", # Unnecessary `else` after `return` statement
    "COM812", # Conflict with formatter
    "SIM108", # Use ternary operator
    "SIM102", # Use single if
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "T201",
]

[tool.ruff.lint.isort]
lines-between-types = 1
known-third-party = [
    "qgis",
]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.flake8-annotations]
ignore-fully-untyped = true
suppress-none-returning = true
suppress-dummy-args = true

#
# Mypy
#
[tool.mypy]
python_version = "3.12"
exclude = [
    "^tests/fixtures/*$",
]

[[tool.mypy.overrides]]
module = [
    "qgis.*",
]
ignore_missing_imports = true

#
# Coverage
#
[tool.coverage.run]
source = ["qgis_plugin_package_ci"]
branch = true
relative_files = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    "if TYPE_CHECKING",
    '\(Protocol\):$',
]

